+++ src/app/components/content-atm-inventories-workflows/list-view.js
@@ -121,18 +121,9 @@ export default Component.extend(I18n, {
     'atmInventory',
     'onOpenAtmWorkflowSchemaRevision',
     function uploadAtmWorkflowSchemaAction() {
-      const {
-        workflowActions,
-        atmInventory,
-        onOpenAtmWorkflowSchemaRevision,
-      } = this.getProperties(
-        'workflowActions',
-        'atmInventory',
-        'onOpenAtmWorkflowSchemaRevision'
-      );
-
-      const action = workflowActions.createUploadAtmWorkflowSchemaAction({
-        atmInventory,
+      const action = this.workflowActions.createUploadAtmRecordAction({
+        atmModelName: 'atmWorkflowSchema',
+        atmInventory: this.atmInventory,
       });
       // After successful workflow schema upload, open it
       action.addExecuteHook(actionResult => {
@@ -142,11 +133,11 @@ export default Component.extend(I18n, {
         } = getProperties(actionResult, 'status', 'result');
         if (status === 'done' && result) {
           const {
-            atmWorkflowSchema,
+            atmRecord,
             revisionNumber,
           } = result;
-          if (atmWorkflowSchema && revisionNumber) {
-            onOpenAtmWorkflowSchemaRevision(atmWorkflowSchema, revisionNumber);
+          if (atmRecord && revisionNumber) {
+            this.onOpenAtmWorkflowSchemaRevision?.(atmRecord, revisionNumber);
           }
         }
       });