+++ src/tests/integration/utils/workflow-actions/dump-atm-lambda-revision-action-test.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import { describe, it, beforeEach } from 'mocha';
 import { setupRenderingTest } from 'ember-mocha';
-import DumpAtmWorkflowSchemaRevisionAction from 'onezone-gui/utils/workflow-actions/dump-atm-workflow-schema-revision-action';
+import DumpAtmLambdaRevisionAction from 'onezone-gui/utils/workflow-actions/dump-atm-lambda-revision-action';
 import sinon from 'sinon';
 import { Promise } from 'rsvp';
 import { lookupService } from '../../../helpers/stub-service';
@@ -9,21 +9,21 @@ import { get, getProperties } from '@ember/object';
 import { settled } from '@ember/test-helpers';
 import { suppressRejections } from '../../../helpers/suppress-rejections';
 
-const atmWorkflowSchemaId = 'wfkId';
+const atmLambdaId = 'lmdId';
 
 describe(
-  'Integration | Utility | workflow-actions/dump-atm-workflow-schema-revision-action',
+  'Integration | Utility | workflow-actions/dump-atm-lambda-revision-action',
   function () {
     setupRenderingTest();
 
     beforeEach(function () {
       const downloadSpy = sinon.spy();
       this.setProperties({
-        action: DumpAtmWorkflowSchemaRevisionAction.create({
+        action: DumpAtmLambdaRevisionAction.create({
           ownerSource: this.owner,
           context: {
-            atmWorkflowSchema: {
-              entityId: atmWorkflowSchemaId,
+            atmLambda: {
+              entityId: atmLambdaId,
             },
             revisionNumber: 3,
           },
@@ -39,30 +39,30 @@ describe(
         icon,
         title,
       } = getProperties(this.get('action'), 'className', 'icon', 'title');
-      expect(className).to.equal('dump-atm-workflow-schema-revision-action-trigger');
+      expect(className).to.equal('dump-atm-lambda-revision-action-trigger');
       expect(icon).to.equal('browser-download');
       expect(String(title)).to.equal('Download (json)');
     });
 
-    it('executes dumping workflow (success scenario)', async function () {
+    it('executes dumping lambda (success scenario)', async function () {
       const dumpStub = sinon
-        .stub(lookupService(this, 'workflow-manager'), 'getAtmWorkflowSchemaDump')
+        .stub(lookupService(this, 'workflow-manager'), 'getAtmLambdaDump')
         .resolves(123);
 
       const actionResultPromise = this.get('action').execute();
       const actionResult = await actionResultPromise;
 
       expect(dumpStub).to.be.calledOnce
-        .and.to.be.calledWith(atmWorkflowSchemaId, 3);
+        .and.to.be.calledWith(atmLambdaId, 3);
       expect(this.get('downloadSpy')).to.be.calledOnce
         .and.to.be.calledWith(sinon.match({ dataString: '123' }));
       expect(get(actionResult, 'status')).to.equal('done');
     });
 
-    it('executes dumping workflow revision (failure scenario)', async function () {
+    it('executes dumping lambda revision (failure scenario)', async function () {
       suppressRejections();
       let rejectDump;
-      sinon.stub(lookupService(this, 'workflow-manager'), 'getAtmWorkflowSchemaDump')
+      sinon.stub(lookupService(this, 'workflow-manager'), 'getAtmLambdaDump')
         .returns(new Promise((resolve, reject) => rejectDump = reject));
       const failureNotifySpy = sinon.spy(
         lookupService(this, 'global-notify'),
@@ -75,7 +75,7 @@ describe(
       const actionResult = await actionResultPromise;
 
       expect(failureNotifySpy).to.be.calledWith(
-        sinon.match.has('string', 'dumping workflow revision'),
+        sinon.match.has('string', 'dumping lambda revision'),
         'someError'
       );
       expect(this.get('downloadSpy')).not.to.be.called;